Moves the CheckResult function to the util package to be re-used else where
