+++ data/src/main/java/me/androidbox/data/remote/util/CheckResult.kt
@@ -1,0 +1,22 @@
+package me.androidbox.data.remote.util
+
+import kotlinx.coroutines.CancellationException
+import retrofit2.HttpException
+
+object CheckResult {
+    suspend fun <T> checkResult(block: suspend () -> T): Result<T> {
+        return try {
+            return Result.success(block())
+        }
+        catch (httpException: HttpException) {
+            Result.failure(httpException)
+        }
+        catch (exception: Exception) {
+            if (exception is CancellationException) {
+                throw exception
+            }
+            Result.failure(exception)
+        }
+    }
+
+}
